REPO-4114 POST /nodes/{nodeId}/children api does not use Transform Se… (#64)
* REPO-4114 POST /nodes/{nodeId}/children api does not use Transform Service for renditions

Changes also to allow multiple renditions to be requested.

The status code 404 is now returned when the rendition id (name) is unknown. Previously this would have been ignored when uploading new content. - Bug fix

The status code 409 is returned if all requested renditions have already been created when requesting new renditions. If some but not all have already been created this code is not returned. 409 was previously returned when the single rendition already existed.